chore: refactor table action bottom sheet
